+++ .github/workflows/backend.yml
@@ -47,7 +47,7 @@ jobs:
         run: |
           tox -e ${{ matrix.toxenv }}
         env: 
-          TEST_RUNNER: cfgov.test.StdoutCapturingTestRunner 
+          TEST_RUNNER: core.testutils.runners.StdoutCapturingTestRunner 
 
       - name: Prepare test coverage
         if: matrix.toxenv == 'unittest-current'
